+++ MidiPanelModule.py
@@ -36,10 +36,10 @@ else:
     from . import PitchUtils
 
 import bpy
-from typing import Callable
-from .NLAMidiCopierModule import NLAMidiCopier, NLAMidiInstrumentCopier, NLAMidiAllInstrumentCopier
+from typing import Callable, Tuple
+from .NLAMidiCopierModule import NLAMidiCopier, NLAMidiInstrumentCopier, NLAMidiAllInstrumentCopier, NLABulkMidiCopier
 from .MidiInstrumentModule import AddInstrument, DeleteInstrument, AddActionToInstrument, RemoveActionFromInstrument, \
-    TransposeInstrument, CopyMidiPanelActionToInstrument
+    TransposeInstrument
 from . import midi_data
 from bpy.props import EnumProperty
 
@@ -117,7 +117,7 @@ class MidiPanel(bpy.types.Panel):
         parent_layout.separator()
 
         col = parent_layout.column(align=True)
-        col.enabled = midi_data.id_type_is_object(note_action_property.id_type)
+        col.enabled = midi_data.can_resolve_data_from_selected_objects(note_action_property.id_type)
         if is_main_property:
             col.prop(note_action_property, "copy_to_selected_objects")
         col.prop(note_action_property, "duplicate_object_on_overlap")
@@ -278,22 +278,89 @@ class CopyToInstrumentPanel(bpy.types.Panel):
     bl_space_type = "NLA_EDITOR"
     bl_region_type = "UI"
     bl_category = "Midi"
-    bl_label = "Copy to Instrument"
+    bl_label = "Copy Along Path and Copy to Instrument"
     bl_idname = "ANIMATION_PT_midi_copy_to_instrument_panel"
     bl_options = {'DEFAULT_CLOSED'}
 
     def draw(self, context):
         col = self.layout.column(align=True)
         midi_data_property = context.scene.midi_data_property
-        col.prop(midi_data_property, "copy_to_instrument_selected_instrument")
-        col.prop(midi_data_property, "copy_to_instrument_selected_note_id")
-        col.prop(midi_data_property, "copy_to_instrument_note_search_string")
+        bulk_copy_property = midi_data_property.bulk_copy_property
+
+        note_column = col.column(align=True)
+        note_column.enabled = not bulk_copy_property.copy_along_path
+        note_column.prop(midi_data_property, "copy_to_instrument_selected_note_id")
+        note_column.prop(midi_data_property, "copy_to_instrument_note_search_string")
+
+        note_action_property = midi_data_property.note_action_property
+
+        copy_along_path_row = col.row()
+        copy_along_path_row.prop(bulk_copy_property, "copy_along_path")
+        # copy along path only works if the id_type corresponds to an object type
+        copy_along_path_enabled = midi_data.can_resolve_data_from_selected_objects(note_action_property.id_type)
+        copy_along_path_row.enabled = copy_along_path_enabled
+
+        if bulk_copy_property.copy_along_path:
+            box = PanelUtils.draw_collapsible_box(col, "Path options", bulk_copy_property,
+                                                  "copy_along_path_options_expanded")[0]
+            if bulk_copy_property.copy_along_path_options_expanded:
+                box_col = box.column(align=True)
+                box_col.prop(bulk_copy_property, "bulk_copy_curve")
+                box_col.prop(bulk_copy_property, "bulk_copy_starting_note")
+                search_row = PanelUtils.indented_row(box_col)
+                search_row.prop(bulk_copy_property, "bulk_copy_starting_note_search_string")
+                box_col.prop(bulk_copy_property, "scale_filter_type")
+                scale_row = PanelUtils.indented_row(box_col)
+                scale_row.enabled = not bulk_copy_property.scale_filter_type == "No filter"
+                scale_row.prop(bulk_copy_property, "scale_filter_scale")
+                box_col.prop(bulk_copy_property, "only_notes_in_selected_track")
+
+        col.prop(bulk_copy_property, "copy_to_instrument")
+        instrument_row = PanelUtils.indented_row(col)
+        instrument_row.enabled = bulk_copy_property.copy_to_instrument
+        instrument_row.prop(midi_data_property, "copy_to_instrument_selected_instrument")
+
         copy_button_row = col.row()
-        copy_button_row.enabled = midi_data_property.copy_to_instrument_selected_instrument is not None and \
-                                  midi_data_property.copy_to_instrument_selected_instrument \
-                                  != midi_data.NO_INSTRUMENT_SELECTED and \
-                                  not midi_data_property.note_action_property.copy_to_selected_objects
-        copy_button_row.operator(CopyMidiPanelActionToInstrument.bl_idname)
+        copy_button_enabled, disabled_tooltip = self.copy_button_enabled(midi_data_property, bulk_copy_property)
+        copy_button_row.enabled = copy_button_enabled
+        animate_button_text = "Copy to Instrument" if bulk_copy_property.copy_to_instrument else "Copy Action to Notes"
+        animate_operator = copy_button_row.operator(NLABulkMidiCopier.bl_idname, text=animate_button_text)
+
+        tooltip = "Copy the action to the selected instrument" if bulk_copy_property.copy_to_instrument else \
+            "Copy the selected Action to the selected note"
+        if disabled_tooltip:
+            tooltip = tooltip + "." + disabled_tooltip
+        animate_operator.tooltip = tooltip
+
+    @staticmethod
+    def copy_button_enabled(midi_data_property, bulk_copy_property) -> Tuple[bool, str]:
+        disabled_tooltip: str = ""  # reason why the copy button is disabled
+        copy_button_enabled = True
+        if bulk_copy_property.copy_to_instrument:
+            print(midi_data_property.copy_to_instrument_selected_instrument)
+            if midi_data_property.copy_to_instrument_selected_instrument is None or \
+                    midi_data_property.copy_to_instrument_selected_instrument == midi_data.NO_INSTRUMENT_SELECTED:
+                copy_button_enabled = False
+                disabled_tooltip += "\n  ! No instrument selected"
+
+            if midi_data_property.note_action_property.copy_to_selected_objects and \
+                    not bulk_copy_property.copy_along_path:
+                copy_button_enabled = False
+                disabled_tooltip += "\n  ! Copy to selected objects (in NLA Midi Panel) not valid for instruments"
+        else:
+            midi_file = midi_data_property.midi_file
+            if midi_file is None or len(midi_file) == 0:
+                copy_button_enabled = False
+                disabled_tooltip += "\n  ! No midi file selected"
+            if midi_data_property.note_action_property.action is None:
+                copy_button_enabled = False
+                disabled_tooltip += "\n  ! No action selected (in NLA Midi Panel)"
+
+        if bulk_copy_property.copy_along_path and bulk_copy_property.bulk_copy_curve is None:
+            copy_button_enabled = False
+            disabled_tooltip += "\n  ! No path selected"
+
+        return copy_button_enabled, disabled_tooltip
 
 
 class MidiSettingsPanel(bpy.types.Panel):